pyproject.toml
[tool.poetry]
name = "data-integration-and-transformation-library"
version = "0.1.0"
description = "A library that aims to simplify etl/elt generation for standard use cases"
authors = [
    {name = "Jochen Gietzen", email = "jochen.gietzen@gmail.com"}
]
license = "MIT"
readme = "README.md"
packages = [{include = "ditl", from="src"}]

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
pydantic = "^2.7.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
